Refactor empresas tab into screen component
